Further separate src from other packages and add readmes and license files
