+++ test/custom-shallow-renderer/Readme.md
@@ -1,0 +1,8 @@
+# test/custom-shallow-renderer
+
+Most/many of tests in this directory are taken from Enzyme's
+`react-shallow-renderer`. Test files that are taken from that repo include a
+comment at the top of the file indicating such. These test files have been
+modified to be TypeScript, formatted to this repo's formatting standards, and
+updated to match which features Preact updates (i.e. tests asserting features
+Preact doesn't support are skipped).
